Add agent module, tests, and update dependencies
Introduced src/backend/agent.py for agent logic and tests/test_agent.py for agent unit tests. Updated FastAPI backend (main.py) and Streamlit frontend (app.py) to integrate the new agent structure. Enhanced README with setup, Windows instructions, and clarified project structure. Added Windows start script (start.ps1) and project contribution guidelines (.github/copilot-instructions.md). Upgraded LangChain and related dependencies in pyproject.toml and poetry.lock for compatibility with latest features and Python 3.13.
