+++ tests/test_main.py
@@ -1,62 +1,9 @@
-import pytest
 from unittest.mock import MagicMock, patch
-from src.backend.main import app, create_agent, Question
+from src.backend.main import app, Question
 from fastapi.testclient import TestClient
 
 # Initialize TestClient for FastAPI application
 client = TestClient(app)
-
-from langchain_core.runnables import Runnable  # noqa: F401
-class MockOllama(Runnable):
-    def __init__(self, *args, **kwargs):
-        pass
-    def invoke(self, *args, **kwargs):
-        return ""
-    def bind(self, *args, **kwargs):
-        return self
-
-def test_create_agent_missing_env(monkeypatch):
-    """Test create_agent raises error if TAVILY_API_KEY is missing."""
-    monkeypatch.delenv("TAVILY_API_KEY", raising=False)
-    from src.backend import main
-    # Patch Ollama to avoid real model loading
-    main.Ollama = MockOllama
-    with pytest.raises(ValueError):
-        main.create_agent()
-
-def test_rag_fusion_retrieval_empty_docs(monkeypatch):
-    """Test rag_fusion_retrieval returns message if no docs found."""
-    from src.backend import main
-    monkeypatch.setenv("TAVILY_API_KEY", "dummy")
-    main.vector_store = MagicMock()
-    main.vector_store.as_retriever.return_value.invoke.return_value = []
-    main.Ollama = MockOllama
-    agent = main.create_agent()
-    tool = next(t for t in agent.tools if t.name == "Local Document Search")
-    assert tool("irrelevant query") == "No relevant information found in local documents."
-
-def test_summarized_web_search_empty_results(monkeypatch):
-    """Test summarized_web_search returns message if no web results found."""
-    from src.backend import main
-    monkeypatch.setenv("TAVILY_API_KEY", "dummy")
-    main.Ollama = MockOllama
-    main.TavilyClient = MagicMock()
-    main.TavilyClient.return_value.search.return_value = {"results": []}
-    agent = main.create_agent()
-    tool = next(t for t in agent.tools if t.name == "Web Search")
-    assert tool("irrelevant query") == "No results found on the web."
-
-
-@patch("src.backend.main.os.getenv", return_value="dummy_tavily_key")
-@patch("src.backend.main.Ollama")
-@patch("src.backend.main.TavilyClient")
-def test_create_agent_initialization(mock_tavily_client, mock_ollama, mock_getenv):
-    """Test if create_agent initializes correctly."""
-    agent_executor = create_agent()
-    assert agent_executor is not None
-    assert hasattr(agent_executor, "agent")
-    assert hasattr(agent_executor, "tools")
-    assert len(agent_executor.tools) == 2 # Local Document Search and Web Search
 
 def test_ask_endpoint_no_agent():
     """Test /ask endpoint when agent is not initialized."""
@@ -71,7 +18,10 @@ def test_ask_endpoint_success(mock_agent_executor):
     mock_agent_executor.invoke.return_value = {"output": "The answer is 42.", "intermediate_steps": []}
     response = client.post("/ask", json={"text": "What is the meaning of life?"})
     assert response.status_code == 200
-    assert response.json() == {"answer": "The answer is 42."}
+    assert response.json() == {
+        "answer": "The answer is 42.",
+        "thinking_steps": []
+    }
     mock_agent_executor.invoke.assert_called_once_with({"input": "What is the meaning of life?"})
 
 @patch("src.backend.main.agent_executor")
@@ -92,9 +42,9 @@ def test_upload_pdf_endpoint_error(mock_create_agent, mock_process_and_store_doc
         f.write(b"dummy pdf content")
     with open(pdf_path, "rb") as f:
         response = client.post("/upload/pdf", files={"file": ("dummy.pdf", f, "application/pdf")})
-    assert response.status_code == 500 or response.status_code == 200
-    # Should return error or handle gracefully
-    mock_load_documents_from_pdf.assert_called_once_with(str(pdf_path))
+    assert response.status_code == 400
+    assert response.json()["detail"] == "PDF load error"
+    mock_load_documents_from_pdf.assert_called_once()
 
 @patch("src.backend.main.load_documents_from_url")
 @patch("src.backend.main.process_and_store_documents")
